# trial use of omp in client/common