* adds openmp to speed up reference herkx calculations