@TorreZuk 2020-03-30
Sure I can produce a time comparision and post shortly, I only verified in the disassembly saw an omp fork.  This is a reference function, no random numbers generated by threads...

Back to Lee's comment  -- presuming your random numbers per thread will likely be sent to either columns or batches it will still result in non-determinism in the input data based on random omp scheduling and total threads available if we parallelize the random data creation process.    #922 has so much in it already not sure if I will decode that part from it.