@TorreZuk 2020-03-30
Timing individual cpu reference locally with small data shows openmp is slower (startup overhead)
./rocblas-bench -f herkx -r f32_c --uplo L --transposeA N -n 199 -k 33 --alpha 1.5 --alphai -1 --lda 199 --ldb 199 --beta 1 --ldc 199 -v 1
~ 8 +/- 6 ms (openmp)
~ 0.3 ms (OLD == current non OpenMP reference)
larger data, then it wins and is where the times are much bigger:
./rocblas-bench -f herkx -r f64_c --uplo L --transposeA C -n 1024 -k 1200 --alpha -2 --alphai 0.5 --lda 1200 --ldb 1200 --beta -1 --ldc 1024 -v 1
~ 88 +/- 6 ms (openmp)
~ 540 +/- 10 ms (OLD)

Testing shows improved timing for large tests and slowdown in small tests but cumulative slow down in small tests isn't significant compared to overall test time reductions.

 PR CI tests for ubuntu18 && gfx900

OpenMP
[----------] 144 tests from quick/herkx
[----------] 144 tests from quick/herkx (158 ms total)
[----------] 130 tests from pre_checkin/herkx
[----------] 130 tests from pre_checkin/herkx (1291 ms total)
[----------] 288 tests from quick/herkx_batched
[----------] 288 tests from quick/herkx_batched (107 ms total)

[----------] 258 tests from pre_checkin/herkx_batched
[----------] 258 tests from pre_checkin/herkx_batched (9931 ms total)
[----------] 288 tests from quick/herkx_strided_batched
[----------] 288 tests from quick/herkx_strided_batched (157 ms total)
[----------] 258 tests from pre_checkin/herkx_strided_batched
[----------] 258 tests from pre_checkin/herkx_strided_batched (9269 ms total)

Old
[----------] 144 tests from quick/herkx
[----------] 144 tests from quick/herkx (102 ms total)
[----------] 130 tests from pre_checkin/herkx
[----------] 130 tests from pre_checkin/herkx (1624 ms total)
[----------] 288 tests from quick/herkx_batched
[----------] 288 tests from quick/herkx_batched (117 ms total)

[----------] 258 tests from pre_checkin/herkx_batched
[----------] 258 tests from pre_checkin/herkx_batched (10302 ms total)
[----------] 288 tests from quick/herkx_strided_batched
[----------] 288 tests from quick/herkx_strided_batched (109 ms total)
[----------] 258 tests from pre_checkin/herkx_strided_batched
[----------] 258 tests from pre_checkin/herkx_strided_batched (9373 ms total)



OpenMP
[----------] 64 tests from nightly/herkx
[----------] 64 tests from nightly/herkx (50944 ms total)
[----------] 32 tests from nightly/herkx_batched
[----------] 32 tests from nightly/herkx_batched (54301 ms total)
[----------] 64 tests from nightly/herkx_strided_batched
[----------] 64 tests from nightly/herkx_strided_batched (98800 ms total)

Old
[----------] 64 tests from nightly/herkx
[----------] 64 tests from nightly/herkx (65689 ms total)
[----------] 32 tests from nightly/herkx_batched
[----------] 32 tests from nightly/herkx_batched (58304 ms total)
[----------] 64 tests from nightly/herkx_strided_batched
[----------] 64 tests from nightly/herkx_strided_batched (128473 ms total)

OpenMP
...
258 tests from pre_checkin/herkx_batched (4352 ms total)
64 tests from nightly/herkx_strided_batched (43731 ms total)
[==========] 1526 tests from 9 test cases ran. (98151 ms total)

No OpenMP
...
258 tests from pre_checkin/herkx_strided_batched (3921 ms total)
64 tests from nightly/herkx_strided_batched (70372 ms total)
[==========] 1526 tests from 9 test cases ran. (146722 ms total)