@leekillough 2020-03-31
> 
> 
> Sure I can produce a time comparision and post shortly, I only verified in the disassembly saw an omp fork. This is a reference function, no random numbers generated by threads...
> 
> Back to Lee's comment -- presuming your random numbers per thread will likely be sent to either columns or batches it will still result in non-determinism in the input data based on random omp scheduling and total threads available if we parallelize the random data creation process. #922 has so much in it already not sure if I will decode that part from it.

The change was mainly done so that the thread-safe logging tests could be written, where multiple threads generate random data at the same time and output to the same file, and then it is verified that they did not interfere with each other, garbling the output.

In the past I've written random number generators, and to make them thread-safe yet independent, the same random number generator is used in each thread, but its state is thread_local, and its seed (starting state) is a pseudorandom hash of the thread's ID. We cannot avoid nondeterminism caused by a random permutation of thread order, but we can ensure that if the same program is run again with the same number of threads, then as long as its result does not depend on the order in which the threads execute, it will generate the same sequences in thread 0, thread 1, thread 2, etc.