@TorreZuk 2020-03-31
Andrew you were right the setup time is significant for a single small data run, but as once those omp worker threads are setup I think the cost is minimal so don't expect we should control omp vs non-omp based on data size and just use them for all cases no matter how small.     I'll go ahead and merge unless someone speaks up shortly as all the rel3.3 failures are unrelated.